requirements.txt
numpy
torch==1.9.0
transformers==3.3.1
tensorboard==2.3.0
gitpython
scipy==1.5.2